[backend]: add a bunch of code for agent's reasoning

+++ game-of-points-be/src/main/java/io/rizvan/beans/actors/CompetingEntity.java
@@ -31,7 +31,7 @@ public class CompetingEntity extends GameEntity {
         return speed * weapon.getSpeedModifier();
     }
 
-    public double getDamage() {
+    public int getDamage() {
         return weapon.getDamage();
     }
